feat(abc:st): support enter trigger filter
